third_party/blink/web_tests/platform/generic/fast/dom/HTMLDocument/active-element-gets-unforcusable-expected.txt

Making a focused element invisible should make it blur.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

===> Setting display:none
Event: blur
PASS document.activeElement is document.body
PASS The focusTarget element lost focus.


===> Setting visibility:hidden
Event: blur
PASS document.activeElement is document.body
PASS The focusTarget element lost focus.


PASS successfullyParsed is true

TEST COMPLETE
